Restructure controllers to have one controller per folder
