+++ src/navigation/path_tracking/path_tracking/controllers/pure_pursuit/controller.py
@@ -5,7 +5,7 @@ from rclpy.impl.rcutils_logger import RcutilsLogger
 from utils.geometry import Path2d, Point2d, Pose2d
 from utils.math import clamp
 
-from .path_tracking_config import PurePursuitConfig
+from .config import PurePursuitConfig
 
 
 class PurePursuitController:
